pyproject.toml
[tool.poetry]
name = "waddi"
version = "0.1.0"
description = ""
authors = ["yomnaafifi <yomnaaafifi162@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.4"
djangorestframework = "^3.15.1"
djangorestframework-simplejwt = "^5.3.1"
markdown = "^3.6"
django-filter = "^24.2"
pillow = "^10.3.0"
psycopg = "^3.1.18"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
